Format date in bank reports.
Feels surprising that this change results in so many variant data classes.
Perhaps there's a nicer way to do it, but I'm changing a LocalDate into a String
in a data class which has multiple parent data classes so I'm not sure how
better I could do it at the moment.