Correct bug in monthly aggregated overview - budget total must come from the budget itself, not only budget values for categories that see actual expenditure.
Again, going a bit fast and dangerous but I plan to come back and tidy up the short cuts that I'm leaving here. It's not an ideal working practice but I actually want to use this app right now and I believe I have a good handle on the shortcomings of this approach.